+++ projects/complete/hatched_venn.py
@@ -8,7 +8,7 @@ def draw_hatched_venn_diagram_horizontal(
     center_y: float,
     offset: float = 0.5,
     line_interval: float = 0.05,
-    angle: int = 45,
+    angle: float = 45,
     radius: float = 1,
 ):
     draw_lined_circle(
@@ -51,7 +51,7 @@ def draw_hatched_venn_diagram_vertical(
     center_y: float,
     offset: float = 0.5,
     line_interval: float = 0.05,
-    angle: int = 45,
+    angle: float = 45,
     radius: float = 1,
 ):
     draw_lined_circle(
